refactor: Rename MovieDetailsResponse package
